6-20  tijiao

+++ src/oss/bean/Users.java
@@ -14,13 +14,13 @@ public class Users {
 	private String userAddress;// 用户地址
 	private Long userStatusID;// 用户账号状态
 	private Long userCredit;// 用户信用分
-	private Long userType;// 用户类型
+	private int userType;// 用户类型
 	public Users() {
 
 	}
 	public Users(Long userID, String userName, String userAccount, String userPwd, String userIdentity,
 			String userPhone, String userEmail, Long userBalance, String userIntroduction, String userAddress,
-			Long userStatusID, Long userCredit, Long userType) {
+			Long userStatusID, Long userCredit, int userType) {
 		super();
 		this.userID = userID;
 		this.userName = userName;
@@ -108,10 +108,10 @@ public class Users {
 	public void setUserCredit(Long userCredit) {
 		this.userCredit = userCredit;
 	}
-	public Long getUserType() {
+	public int getUserType() {
 		return userType;
 	}
-	public void setUserType(Long userType) {
+	public void setUserType(int userType) {
 		this.userType = userType;
 	}
